pyproject.toml
[tool.poetry]
package-mode = false
name = "firstapi"
version = "0.1.0"
description = ""
authors = ["Aiven-gwt <v_bia_v@mail.ru>"]
license = "GWT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
pydantic = {extras = ["email"], version = "^2.8.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.34"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
black = "^24.8.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
